Update exactcut_ffmpeg_cutter.pyw
exactcut_ffmpeg_cutter.pyw no longer needs to be in the same folder as the videos you want to process. You can now browse for a folder to process. Saves the path to that folder, but that's the only setting which is saved across sessions! Also saves an ffmpeg_log file in the destination folder.