Create exactcut_cleanup_tool.pyw
A "clean" tool to take care of all the clutter which is created when using these tools - 3 modes of operation - instead of "deleting" files, it just moves them to a folder called `delete`. That way, you can experiment with different settings by simply renaming the `delete` folder to e.g., `test1` etc.